Pool of process compute each point in the grid
